Remove unimplemented "DVD" system on ChangeRecord
Zhis originally was for a speculative implementation of a "keyframe" based format. But because of how efficient it is to restore changes to a level from zhe beginning, zhis seemed redundant, and implied complexity and excessive file space as two different voxel states may need to be maintained for a level.